Add files via upload
Conclusions:

Information about the dataset:

It is a imbalanced dataset and 52% of the data are for patients below 52.

Insights from age of patients:

If age is less than 35 we can say that the patient can live 5 years or longer for sure.When age is less than 46 and above 30the patient may live more than 5 years with error in prediction present as the box plots are overlapping and the amount of overlapping is 25 percent of 81 which is 20. It is found that people with age between 53 and 58 have high probability of living for more than 5 years In other range of ages we can give only probalistic values and they are neither high nor low.

Insights from no of axil_det_nodes:

As axil_nodes_det is in the range of 0-5,the patient will live for more than 5 years once the it crossed 10,the patient may not live for mare than 5 years. It is noted from data that 66.5% of people have less than 3 axil_nodes_dent. 0.05 percent of people have highest number of 46 axil_nodes_dent. They are able to survive. There is 75 percentile probability that when axil_nodes_det are less than 3 the patient will live for more than 5 years , of that there is a 50 percent chance of commiting error. There are many outliers so we should be very careful 90 % of people having less than 8 axil_nodes_det survive for more than 5 years Chances of surving for more than 5 increases if the axil_nodes_det is near to zero